# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "IamController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page number, starts from 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size, max 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Fuzzy search keyword",
            "schema": {
              "example": "demo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List admin users",
        "tags": [
          "admin-iam"
        ]
      },
      "post": {
        "operationId": "IamController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create admin user",
        "tags": [
          "admin-iam"
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "IamController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get admin user detail",
        "tags": [
          "admin-iam"
        ]
      },
      "patch": {
        "operationId": "IamController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update admin user",
        "tags": [
          "admin-iam"
        ]
      }
    },
    "/api/v1/admin/users/{id}/roles": {
      "post": {
        "operationId": "IamController_bindUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindUserRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bind roles to admin user",
        "tags": [
          "admin-iam"
        ]
      }
    },
    "/api/v1/admin/roles": {
      "get": {
        "operationId": "IamController_listRoles",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page number, starts from 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size, max 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Fuzzy search keyword",
            "schema": {
              "example": "demo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List roles",
        "tags": [
          "admin-iam"
        ]
      },
      "post": {
        "operationId": "IamController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create role",
        "tags": [
          "admin-iam"
        ]
      }
    },
    "/api/v1/admin/roles/{id}": {
      "get": {
        "operationId": "IamController_getRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get role detail",
        "tags": [
          "admin-iam"
        ]
      },
      "patch": {
        "operationId": "IamController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update role",
        "tags": [
          "admin-iam"
        ]
      }
    },
    "/api/v1/admin/roles/{id}/permissions": {
      "patch": {
        "operationId": "IamController_assignRolePermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign role permissions",
        "tags": [
          "admin-iam"
        ]
      }
    },
    "/api/v1/admin/permissions": {
      "get": {
        "operationId": "IamController_listPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List permissions",
        "tags": [
          "admin-iam"
        ]
      }
    },
...
  },
  "info": {
    "title": "rtnn backend api",
    "description": "Template backend kernel for RTNN",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "AdminUserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "acc_01JABCD123"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "name": {
            "type": "string",
            "example": "Operations Admin"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "locked"
            ],
            "example": "active"
          },
          "roles": {
            "example": [
              "Operations Admin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roleIds": {
            "example": [
              "role_01JABCD123"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastLoginAt": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "status",
          "roles",
          "lastLoginAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminUserListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "RoleSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "role_01JABCD123"
          },
          "slug": {
            "type": "string",
            "example": "ops-admin"
          },
          "code": {
            "type": "string",
            "example": "OPS_ADMIN"
          },
          "name": {
            "type": "string",
            "example": "Operations Admin"
          },
          "description": {
            "type": "object",
            "example": "Can operate day-to-day admin workflows"
          },
          "tenantId": {
            "type": "object",
            "example": null
          },
          "permissionKeys": {
            "example": [
              "admin:users:view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "example": [
              "admin:users:view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminUserDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "acc_01JABCD123"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "name": {
            "type": "string",
            "example": "Operations Admin"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "locked"
            ],
            "example": "active"
          },
          "roles": {
            "example": [
              "Operations Admin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roleIds": {
            "example": [
              "role_01JABCD123"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastLoginAt": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "permissions": {
            "example": [
              "admin:users:view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rolesDetailed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "status",
          "roles",
          "lastLoginAt",
          "createdAt",
          "updatedAt",
          "permissions",
          "rolesDetailed"
        ]
      },
      "CreateAdminUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ops-admin@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "StrongPass123!"
          },
          "displayName": {
            "type": "string",
            "example": "Operations Admin"
          },
          "name": {
            "type": "string",
            "example": "Operations Admin"
          },
          "roleIds": {
            "example": [
              "role_01"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "locked"
            ],
            "example": "active"
          },
          "tenantId": {
            "type": "string",
            "example": "default"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "Operations Admin"
          },
          "name": {
            "type": "string",
            "example": "Operations Admin"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "StrongPass123!"
          },
          "roleIds": {
            "example": [
              "role_01"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "locked"
            ],
            "example": "active"
          }
        }
      },
      "BindUserRolesDto": {
        "type": "object",
        "properties": {
          "roleSlugs": {
            "example": [
              "super-admin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roleIds": {
            "example": [
              "role_01"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RoleListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "ops-manager"
          },
          "name": {
            "type": "string",
            "example": "Ops Manager"
          },
          "description": {
            "type": "string",
            "example": "Operations role"
          },
          "permissionKeys": {
            "example": [
              "admin:dashboard:view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "ops-manager"
          },
          "name": {
            "type": "string",
            "example": "Ops Manager"
          },
          "description": {
            "type": "string",
            "example": "Operations role"
          },
          "permissionKeys": {
            "example": [
              "admin:dashboard:view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AssignRolePermissionsDto": {
        "type": "object",
        "properties": {
          "permissionKeys": {
            "example": [
              "admin:dashboard:view",
              "admin:users:view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissionKeys"
        ]
      },
      "PermissionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "perm_01JABCD123"
          },
          "key": {
            "type": "string",
            "example": "admin:users:view"
          },
          "name": {
            "type": "string",
            "example": "View admin users"
          },
          "description": {
            "type": "object",
            "example": "Allows reading admin user profiles"
          }
        },
        "required": [
          "id",
          "key",
          "name"
        ]
      },
...
    }
  }
}